Hand-written version of returning simple structs from a dart service.
Variable size stuff not yet supported.

R=kasperl@google.com

Review URL: https://codereview.chromium.org/888563004

+++ tests/service_tests/person/cc/struct.h
@@ -7,6 +7,8 @@
 
 #include "include/service_api.h"
 
+#include <inttypes.h>
+
 class Builder;
 class MessageBuilder;
 
@@ -100,7 +102,7 @@ class Builder {
   BuilderSegment* segment() const { return segment_; }
   int offset() const { return offset_; }
 
-  int InvokeMethod(ServiceId service, MethodId method);
+  int64_t InvokeMethod(ServiceId service, MethodId method);
 
  protected:
   Builder(BuilderSegment* segment, int offset)
